chore(ui): decompose grid-gallery and language-selector
