+++ src/components/ui/language-selector.tsx
@@ -1,7 +1,8 @@
+import { useDropdown } from "@components/hooks/use-dropdown";
 import { type Lang, languages } from "@i18n/locale";
 import { getCanonicalBasePath, getTranslatedPath } from "@i18n/path";
 import { cn } from "@lib/cn";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback } from "react";
 import { LuGlobe } from "react-icons/lu";
 
 export interface LanguageSelectorProps {
@@ -14,130 +15,58 @@ export interface LanguageSelectorProps {
 
 const langEntries = Object.entries(languages) as [Lang, string][];
 
-function getMenuLinks(menuRef: React.RefObject<HTMLDivElement | null>) {
-  return Array.from(menuRef.current?.querySelectorAll("a") ?? []);
-}
-
-function focusMenuItemByOffset(
-  menuRef: React.RefObject<HTMLDivElement | null>,
-  offset: number
-) {
-  const links = getMenuLinks(menuRef);
-  const currentIndex = links.indexOf(
-    document.activeElement as HTMLAnchorElement
-  );
-  const nextIndex =
-    offset > 0
-      ? (currentIndex + 1) % links.length
-      : (currentIndex - 1 + links.length) % links.length;
-  links[nextIndex]?.focus();
-}
-
-export function LanguageSelector({
+function InlineLanguageSelector({
   lang,
-  variant = "dropdown",
-  currentPath,
-  className = "",
-  "data-testid": dataTestId = "language-selector",
-}: Readonly<LanguageSelectorProps>) {
-  const [isOpen, setIsOpen] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const buttonRef = useRef<HTMLButtonElement>(null);
-
-  const getBasePath = useCallback(() => {
-    if (currentPath) {
-      return currentPath;
-    }
-    if (typeof window === "undefined") {
-      return "/";
-    }
-    return getCanonicalBasePath(new URL(window.location.href));
-  }, [currentPath]);
-
-  const getHref = useCallback(
-    (targetLang: Lang) => {
-      const basePath = getBasePath();
-      const tp = getTranslatedPath(targetLang);
-      return tp(basePath, targetLang);
-    },
-    [getBasePath]
+  getHref,
+  className,
+  dataTestId,
+}: {
+  lang: Lang;
+  getHref: (code: Lang) => string;
+  className: string;
+  dataTestId: string;
+}) {
+  return (
+    <div className={cn("flex space-x-4", className)} data-testid={dataTestId}>
+      {langEntries.map(([code]) => (
+        <a
+          className={cn(
+            "rounded px-3 py-1",
+            code === lang
+              ? "bg-accent text-accent-foreground"
+              : "hover:bg-muted"
+          )}
+          data-lang={code}
+          href={getHref(code)}
+          key={code}
+        >
+          {code.toUpperCase()}
+        </a>
+      ))}
+    </div>
   );
+}
 
-  // Close on outside click
-  useEffect(() => {
-    if (!isOpen) {
-      return;
-    }
-
-    function handleClickOutside(e: MouseEvent) {
-      const target = e.target as Node;
-      if (containerRef.current && !containerRef.current.contains(target)) {
-        setIsOpen(false);
-      }
-    }
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, [isOpen]);
-
-  function handleButtonKeyDown(e: React.KeyboardEvent) {
-    if (isOpen) {
-      handleOpenKeyDown(e);
-    } else {
-      handleClosedKeyDown(e);
-    }
-  }
-
-  function handleClosedKeyDown(e: React.KeyboardEvent) {
-    if (e.key === "ArrowDown" || e.key === "Enter" || e.key === " ") {
-      e.preventDefault();
-      setIsOpen(true);
-      setTimeout(() => {
-        const firstLink = menuRef.current?.querySelector("a");
-        firstLink?.focus();
-      }, 0);
-    }
-  }
-
-  function handleOpenKeyDown(e: React.KeyboardEvent) {
-    if (e.key === "Escape") {
-      e.preventDefault();
-      setIsOpen(false);
-      buttonRef.current?.focus();
-      return;
-    }
-
-    if (e.key === "ArrowDown") {
-      e.preventDefault();
-      focusMenuItemByOffset(menuRef, 1);
-    } else if (e.key === "ArrowUp") {
-      e.preventDefault();
-      focusMenuItemByOffset(menuRef, -1);
-    }
-  }
-
-  if (variant === "inline") {
-    return (
-      <div className={cn("flex space-x-4", className)} data-testid={dataTestId}>
-        {langEntries.map(([code]) => (
-          <a
-            className={cn(
-              "rounded px-3 py-1",
-              code === lang
-                ? "bg-accent text-accent-foreground"
-                : "hover:bg-muted"
-            )}
-            data-lang={code}
-            href={getHref(code)}
-            key={code}
-          >
-            {code.toUpperCase()}
-          </a>
-        ))}
-      </div>
-    );
-  }
+function DropdownLanguageSelector({
+  lang,
+  getHref,
+  className,
+  dataTestId,
+}: {
+  lang: Lang;
+  getHref: (code: Lang) => string;
+  className: string;
+  dataTestId: string;
+}) {
+  const {
+    isOpen,
+    toggle,
+    containerRef,
+    menuRef,
+    buttonRef,
+    handleButtonKeyDown,
+    handleMenuKeyDown,
+  } = useDropdown();
 
   return (
     <div
@@ -149,7 +78,7 @@ export function LanguageSelector({
         aria-expanded={isOpen}
         aria-haspopup="listbox"
         className="flex items-center space-x-2 rounded-lg px-3 py-2 transition-colors hover:bg-muted"
-        onClick={() => setIsOpen(!isOpen)}
+        onClick={toggle}
         onKeyDown={handleButtonKeyDown}
         ref={buttonRef}
         type="button"
@@ -161,7 +90,7 @@ export function LanguageSelector({
       {isOpen && (
         <div
           className="absolute right-0 mt-2 w-32 overflow-hidden rounded-lg border border-border bg-card shadow-lg"
-          onKeyDown={handleOpenKeyDown}
+          onKeyDown={handleMenuKeyDown}
           ref={menuRef}
           role="listbox"
         >
@@ -184,5 +113,46 @@ export function LanguageSelector({
         </div>
       )}
     </div>
+  );
+}
+
+export function LanguageSelector({
+  lang,
+  variant = "dropdown",
+  currentPath,
+  className = "",
+  "data-testid": dataTestId = "language-selector",
+}: Readonly<LanguageSelectorProps>) {
+  const getHref = useCallback(
+    (targetLang: Lang) => {
+      const basePath =
+        currentPath ??
+        (typeof window === "undefined"
+          ? "/"
+          : getCanonicalBasePath(new URL(window.location.href)));
+      const tp = getTranslatedPath(targetLang);
+      return tp(basePath, targetLang);
+    },
+    [currentPath]
+  );
+
+  if (variant === "inline") {
+    return (
+      <InlineLanguageSelector
+        className={className}
+        dataTestId={dataTestId}
+        getHref={getHref}
+        lang={lang}
+      />
+    );
+  }
+
+  return (
+    <DropdownLanguageSelector
+      className={className}
+      dataTestId={dataTestId}
+      getHref={getHref}
+      lang={lang}
+    />
   );
 }
